requirements.txt
tika
parsr-client
pandas
wget
aspose-words
python-magic